+++ pulsar/apps/test/utils.py
@@ -43,7 +43,6 @@ import unittest
 from inspect import isclass
 from functools import partial
 from contextlib import contextmanager
-from asyncio import Future
 
 try:
     from unittest.case import _ExpectedFailure as ExpectedFailure
@@ -53,7 +52,7 @@ except ImportError:
 import pulsar
 from pulsar import (get_actor, send, multi_async, async, future_timeout,
                     TcpServer, coroutine_return, new_event_loop,
-                    format_traceback)
+                    Future, format_traceback)
 from pulsar.async.proxy import ActorProxyFuture
 from pulsar.utils.importer import module_attribute
 from pulsar.apps.data import create_store